Refactor code to use dependency injection starting from MediaSessionHandlerAdapter. That way we can use interfaces later to inject stubs when implementing unit tests
