+++ app/src/main/java/com/fivegmag/a5gmsmediastreamhandler/player/IQoeMetricsReporter.kt
@@ -2,7 +2,7 @@ package com.fivegmag.a5gmsmediastreamhandler.player
 
 import com.fivegmag.a5gmscommonlibrary.qoeMetricsReporting.QoeMetricsRequest
 
-interface QoeMetricsReporter {
+interface IQoeMetricsReporter {
 
     fun initialize(lastQoeMetricsRequest: QoeMetricsRequest)
 
@@ -11,4 +11,6 @@ interface QoeMetricsReporter {
     fun reset()
 
     fun resetState()
+
+    fun setLastQoeMetricsRequest(lastQoeMetricsRequest: QoeMetricsRequest)
 }
